refactor: 收敛工具域路由边界
