+++ qq-maid-core/src/runtime/tools/todo/mod.rs
@@ -17,6 +17,7 @@ pub(crate) mod edit_patch;
 pub(crate) mod flow;
 pub(crate) mod format;
 mod freshness;
+pub(crate) mod interaction_state;
 mod json;
 pub(crate) mod ops;
 pub(crate) mod pending;
